Move entity to shared kernel
