--- src/Urleso.Domain/Primitives/EntityId.cs
@@ -1,3 +1,0 @@
-﻿namespace Urleso.Domain.Primitives;
-
-public abstract record EntityId(Guid Value);
