+++ src/Urleso.SharedKernel/Entity.cs
@@ -1,0 +1,26 @@
+﻿namespace Urleso.SharedKernel;
+
+public abstract class Entity<TEntityId>(TEntityId id) : IEquatable<Entity<TEntityId>>
+    where TEntityId : IEquatable<TEntityId>
+{
+    public TEntityId Id => id;
+
+    public static bool operator !=(Entity<TEntityId>? first, Entity<TEntityId>? second) => !(first == second);
+
+    public static bool operator ==(Entity<TEntityId>? first, Entity<TEntityId>? second) =>
+        first?.Equals(second) ?? ReferenceEquals(first, second);
+
+    public override int GetHashCode() => Id.GetHashCode();
+
+    public override bool Equals(object? obj) => Equals(obj as Entity<TEntityId>);
+
+    public bool Equals(Entity<TEntityId>? other)
+    {
+        if (other is null)
+        {
+            return false;
+        }
+
+        return other.GetType() == GetType() && Id.Equals(other.Id);
+    }
+}
