--- tests/Urleso.Domain.UnitTests/Primitives/DummyEntity.cs
@@ -1,5 +1,0 @@
-﻿using Urleso.Domain.Primitives;
-
-namespace Urleso.Domain.UnitTests.Primitives;
-
-internal sealed class DummyEntity(DummyEntityId id) : Entity<DummyEntityId>(id);
