--- tests/Urleso.Domain.UnitTests/Primitives/DummyEntityId.cs
@@ -1,5 +1,0 @@
-﻿using Urleso.Domain.Primitives;
-
-namespace Urleso.Domain.UnitTests.Primitives;
-
-internal sealed record DummyEntityId(Guid Value) : EntityId(Value);
