+++ tests/Urleso.SharedKernel.UnitTests/DummyEntity.cs
@@ -1,0 +1,3 @@
+﻿namespace Urleso.SharedKernel.UnitTests;
+
+internal sealed class DummyEntity(DummyEntityId id) : Entity<DummyEntityId>(id);
