+++ tests/Urleso.SharedKernel.UnitTests/DummyEntityId.cs
@@ -1,0 +1,3 @@
+﻿namespace Urleso.SharedKernel.UnitTests;
+
+internal sealed record DummyEntityId(Guid Value);
